Add function to get latest image version from Compute Gallery
